[review-fix 4] Move Dynamic Tests helper out of pkg/config/setup
Change:
- Description: Move ConnectionDynamicTestsEnabled to pkg/networkpath/enablement and update collector, reverse DNS, and traceroute callers
- Motivation: comp/ cannot import pkg/config/setup; that import failed cross-OS lint
- Source: lint_cross_windows-x64 lint_cross_macos lint_cross_aix-ppc64
