+++ comp/networkpath/npcollector/impl/BUILD.bazel
@@ -33,10 +33,10 @@ go_library(
         "//comp/rdnsquerier/impl-none",
         "//comp/remote-config/rcclient/types",
         "//pkg/config/remote/data",
-        "//pkg/config/setup",
         "//pkg/config/structure",
         "//pkg/logs/message",
         "//pkg/network/tracer/networkfilter",
+        "//pkg/networkpath/enablement",
         "//pkg/networkpath/payload",
         "//pkg/networkpath/traceroute/config",
         "//pkg/remoteconfig/state",
