+++ pkg/config/setup/BUILD.bazel
@@ -45,7 +45,6 @@ go_library(
         "config_windows.go",
         "constants_compat.go",
         "fixup_init.go",
-        "network_path.go",
         "otlp.go",
         "privateactionrunner.go",
         "process.go",
@@ -88,7 +87,6 @@ dd_agent_go_test(
         "config_init_test.go",
         "config_secret_test.go",
         "config_test.go",
-        "network_path_test.go",
         "privateactionrunner_test.go",
         "process_test.go",
         "system_probe_test.go",
