+++ pkg/networkpath/enablement/BUILD.bazel
@@ -1,0 +1,20 @@
+load("@rules_go//go:def.bzl", "go_library")
+load("//bazel/rules/go:dd_agent_go_test.bzl", "dd_agent_go_test")
+
+go_library(
+    name = "enablement",
+    srcs = ["enablement.go"],
+    importpath = "github.com/DataDog/datadog-agent/pkg/networkpath/enablement",
+    visibility = ["//visibility:public"],
+    deps = ["//pkg/config/model"],
+)
+
+dd_agent_go_test(
+    name = "enablement_test",
+    srcs = ["enablement_test.go"],
+    embed = [":enablement"],
+    deps = [
+        "//pkg/config/mock",
+        "@com_github_stretchr_testify//assert",
+    ],
+)
